Move AssetCommandForm from assets to mission/forms.py
AssetCommandForm operates on mission.models.AssetCommand/MissionAsset;
it belongs in mission alongside those models. Delete the now-empty
assets/forms.py and the orphaned asset-command-form.html template.
