--- assets/templates/asset-command-form.html
@@ -1,3 +1,0 @@
-<table>
-{{ form.as_table }}
-</table>
