+++ mission/forms.py
@@ -3,8 +3,10 @@ Forms for missions
 """
 from django.forms import ModelForm
 
+from assets.models import Asset
+
 from mission.helpers import get_my_assets_not_in_mission
-from .models import Mission, MissionUser, MissionAsset, MissionOrganization
+from .models import AssetCommand, Mission, MissionUser, MissionAsset, MissionOrganization
 
 
 class MissionForm(ModelForm):
@@ -48,3 +50,22 @@ class MissionOrganizationForm(ModelForm):
     class Meta:
         model = MissionOrganization
         fields = ['organization']
+
+
+class AssetCommandForm(ModelForm):
+    """
+    Form for letting user set the next command for an asset.
+    The UI presenting this needs to provide a way to select the position
+    (when the command is one that requires a position)
+    """
+    def __init__(self, *args, **kwargs):
+        self.mission = kwargs.pop('mission')
+        super().__init__(*args, **kwargs)
+        mission_assets = MissionAsset.objects.filter(mission=self.mission, removed__isnull=True)
+        self.fields['asset'].queryset = Asset.objects.filter(
+            pk__in=[mission_asset.asset.pk for mission_asset in mission_assets]
+        )
+
+    class Meta:
+        model = AssetCommand
+        fields = ['asset', 'command', 'reason']
